// backend/src/shared/types/board-permissions.enum.ts
export enum BoardPermission {
    MANAGE_BOARD = 'MANAGE_BOARD',
    DELETE_BOARD = 'DELETE_BOARD',

    MANAGE_ROLES = 'MANAGE_ROLES',
    INVITE_USERS = 'INVITE_USERS',
    KICK_USERS = 'KICK_USERS',

    CREATE_COLUMN = 'CREATE_COLUMN',
    UPDATE_COLUMN = 'UPDATE_COLUMN',
    DELETE_COLUMN = 'DELETE_COLUMN',

    CREATE_CARD = 'CREATE_CARD',
    UPDATE_CARD = 'UPDATE_CARD',
    DELETE_CARD = 'DELETE_CARD',
}